+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Omnibus Ruby CHANGELOG
 Unreleased
 ----------
 - Make build commands output during `log.info` instead of `log.debug`
+- Refactor Chef Sugar into an includable module, permitting DSL methods in both Software and Project definitions
 
 
 v3.1.1 (May 20, 2014)